+++ MuckTimeModifier/TimeModifierPatch.cs
@@ -4,16 +4,18 @@ using HarmonyLib;
 
 public class TimeModifierPatch
 {
+	//public static float ApproxUnitTimeInSeconds = 0f;
 	[HarmonyPatch(typeof(DayCycle), "Awake")]
 	[HarmonyPostfix]
 	private static void SetTimeSpeedAndNightDuration(DayCycle __instance)
 	{
-		float newNightDuration = __instance.nightDuration * Plugin.NightLengthMultiplier;
+		float newNightDuration = Plugin.NightTimeSpeedDivisor;
 		float newTimeSpeed = __instance.timeSpeed / Plugin.TimeSpeedDivisor;
-		Plugin.Log.LogInfo(string.Concat("Changing timeSpeed: ", __instance.timeSpeed, " / ", Plugin.TimeSpeedDivisor, " = ", newTimeSpeed));
+		Plugin.Log.LogInfo(string.Concat("Changing Time Speed: ", __instance.timeSpeed, " / ", Plugin.TimeSpeedDivisor, " = ", newTimeSpeed));
 		__instance.timeSpeed = newTimeSpeed;
-		Plugin.Log.LogInfo(string.Concat("Changing nightDuration: ", __instance.nightDuration, " * ", Plugin.NightLengthMultiplier, " = ", newNightDuration));
+		Plugin.Log.LogInfo(string.Concat("Changing Nighttime Speed divisor: ", __instance.nightDuration, " to ", newNightDuration));
 		__instance.nightDuration = newNightDuration;
+		//ApproxUnitTimeInSeconds = newTimeSpeed;
 	}
 	[HarmonyPatch(typeof(GameManager), "Start")]
 	[HarmonyPostfix]
@@ -24,5 +26,11 @@ public class TimeModifierPatch
 		float newDayDuration = DayCycle.dayDuration * Plugin.DayLengthMultiplier;
 		Plugin.Log.LogInfo(string.Concat("Changing dayDuration: ", DayCycle.dayDuration, " * ", Plugin.DayLengthMultiplier, " = ", newDayDuration));
 		DayCycle.dayDuration = newDayDuration;
+		float dayPortion = newDayDuration / 2;
+		float nightPortion = dayPortion * Plugin.NightTimeSpeedDivisor;
+		float totalPortion = dayPortion + nightPortion;
+
+		//Plugin.Log.LogInfo("Total day duration in seconds: " + newDayDuration / ApproxUnitTimeInSeconds);
+		Plugin.Log.LogInfo(string.Concat("One day cycle is now ", newDayDuration, " units long. Taking into account that nighttime lasts ", Plugin.NightTimeSpeedDivisor, ", the actual duration is ", totalPortion, " units. Daytime is ", dayPortion, " long, and thus takes up ", dayPortion / totalPortion * 100f, "% of the time, and night is ", nightPortion, " long, and thus takes up ", nightPortion / totalPortion * 100f, "% of the time"));
 	}
 }