+++ MuckTimeModifier/CHANGELOG.md
@@ -6,3 +6,8 @@
 
 # v1.2.0
 - Added the ability to configure how frequently Boss Days occur
+
+# v1.3.0
+- Fixed comments explaining what things do now that I correctly understand how time passes in Muck
+- Fixed the behaviour of nighttime time speed. NightLengthMultiplier removed, it's now NightTimeSpeedDivisor, because what Muck does is change the speed at which time passes at nighttime.
+- Information about the length of daytime and nighttime is written to the console on starting a game.